Add files via upload
Uses the Uniswap v2 router contract to do buy and sells. It also does token approvals to the uniswap contract based on settings. All wallet usage for the program is kept in this file and is open source so you can review and feel comfortable that your private keys aren't going to be used anywhere else. Potentially could use other DEX contracts with no or minor modifications.